Use floats for cpu (#410)
* use floats for cpu

* change CPUStats.NumCPU

* missed a return

+++ utils/cpu.go
@@ -14,7 +14,7 @@ import (
 
 type platformCPUMonitor interface {
 	getCPUIdle() (float64, error)
-	numCPU() int
+	numCPU() float64
 }
 
 type CPUStats struct {
@@ -48,7 +48,7 @@ func (c *CPUStats) GetCPUIdle() float64 {
 	return c.idleCPUs.Load()
 }
 
-func (c *CPUStats) NumCPU() int {
+func (c *CPUStats) NumCPU() float64 {
 	return c.platform.numCPU()
 }
 
@@ -72,7 +72,7 @@ func (c *CPUStats) monitorCPULoad() {
 			}
 
 			c.idleCPUs.Store(idle)
-			idleRatio := idle / float64(c.platform.numCPU())
+			idleRatio := idle / c.platform.numCPU()
 
 			if idleRatio < 0.1 {
 				c.warningThrottle(func() { logger.Infow("high cpu load", "load", 1-idleRatio) })
